Fix warnings. Undisable dos2unix
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@5405 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ alpertron64/quad.cc
@@ -230,7 +230,7 @@ struct Quad {
     }
 
     CallbackQuadModElliptic(value, modulus, e);
-    return solutions.points.size() < expected_solutions;
+    return (int)solutions.points.size() < expected_solutions;
   }
 
   // Solve congruence an^2 + bn + c = 0 (mod m) where m is different from zero.
@@ -419,7 +419,7 @@ struct Quad {
     for (;;) {
 
       // Once we have all the solutions, stop looking.
-      if (solutions.points.size() == expected_solutions)
+      if ((int)solutions.points.size() == expected_solutions)
         break;
 
       // This code maintains the factor list as it computes different k.
@@ -705,7 +705,7 @@ Solutions SolveQuad(uint64_t f,
   const int expected_solutions = ChaiWahWuFromFactors(f, factors);
   std::unique_ptr<Quad> quad(new Quad(expected_solutions));
   quad->SolveQuad(f, factors);
-  CHECK(quad->solutions.points.size() == expected_solutions) <<
+  CHECK((int)quad->solutions.points.size() == expected_solutions) <<
     "Sum: " << f <<
     "\nexpected: " << expected_solutions <<
     "\ngot: " << quad->solutions.points.size();
